Update CartController.php
Hello,

there is a bug when entering "0" quantity to an item in the cart. The system will crash and send error message "Devision by Zero is not allowed".

I solved the problem by editing the updateAction function (line 114).

Maybe it's not the smartes solution, but it works. Maybe you can write a better way to solve it. 

Best regards